+++ changelog/README.md
@@ -1,7 +1,7 @@
 ## Changelog
 
 ### 1.0.0-alpha3 (in progress)
-
+- [bug] KAF-39: Partitions don't rebalance in distributed mode across multiple workers
 
 ### 1.0.0-alpha2
 - [improvement] KAF-38: Improve performance by parallelizing record mapping and using partition-key-based batches when possible
